Bug 793201 - Fix compiler warnings in Azure with MSVC r=bas
UltraBlame original commit: 96ef3b8bd9ed911a42f1c27cd23604edc846cda1

+++ gfx/2d/PathHelpers.h
@@ -16,67 +16,67 @@ template <typename T>
 void ArcToBezier(T* aSink, const Point &aOrigin, float aRadius, float aStartAngle,
                  float aEndAngle, bool aAntiClockwise)
 {
-  Point startPoint(aOrigin.x + cos(aStartAngle) * aRadius, 
-                   aOrigin.y + sin(aStartAngle) * aRadius); 
-  
-  aSink->LineTo(startPoint); 
+  Point startPoint(aOrigin.x + cos(aStartAngle) * aRadius,
+                   aOrigin.y + sin(aStartAngle) * aRadius);
+
+  aSink->LineTo(startPoint);
 
   
   
   if (!aAntiClockwise && (aEndAngle < aStartAngle)) {
-    Float correction = ceil((aStartAngle - aEndAngle) / (2.0f * M_PI));
-    aEndAngle += correction * 2.0f * M_PI;
+    Float correction = Float(ceil((aStartAngle - aEndAngle) / (2.0f * M_PI)));
+    aEndAngle += float(correction * 2.0f * M_PI);
   } else if (aAntiClockwise && (aStartAngle < aEndAngle)) {
-    Float correction = ceil((aEndAngle - aStartAngle) / (2.0f * M_PI));
-    aStartAngle += correction * 2.0f * M_PI;
+    Float correction = (Float)ceil((aEndAngle - aStartAngle) / (2.0f * M_PI));
+    aStartAngle += float(correction * 2.0f * M_PI);
   }
-  
+
   
   if (!aAntiClockwise && (aEndAngle - aStartAngle > 2 * M_PI)) {
-    aEndAngle = aStartAngle + 2.0f * M_PI;
+    aEndAngle = float(aStartAngle + 2.0f * M_PI);
   } else if (aAntiClockwise && (aStartAngle - aEndAngle > 2.0f * M_PI)) {
-    aEndAngle = aStartAngle - 2.0f * M_PI;
+    aEndAngle = float(aStartAngle - 2.0f * M_PI);
   }
 
   
   Float arcSweepLeft = fabs(aEndAngle - aStartAngle);
-  
+
   Float sweepDirection = aAntiClockwise ? -1.0f : 1.0f;
-  
+
   Float currentStartAngle = aStartAngle;
-  
+
   while (arcSweepLeft > 0) {
     
     
     Float currentEndAngle;
-    
+
     if (arcSweepLeft > M_PI / 2.0f) {
-      currentEndAngle = currentStartAngle + M_PI / 2.0f * sweepDirection;                                                               
+      currentEndAngle = Float(currentStartAngle + M_PI / 2.0f * sweepDirection);
     } else {
       currentEndAngle = currentStartAngle + arcSweepLeft * sweepDirection;
     }
-    
+
     Point currentStartPoint(aOrigin.x + cos(currentStartAngle) * aRadius,
                             aOrigin.y + sin(currentStartAngle) * aRadius);
     Point currentEndPoint(aOrigin.x + cos(currentEndAngle) * aRadius,
                           aOrigin.y + sin(currentEndAngle) * aRadius);
-    
+
     
     
     
     Float kappa = (4.0f / 3.0f) * tan((currentEndAngle - currentStartAngle) / 4.0f) * aRadius;
-    
+
     Point tangentStart(-sin(currentStartAngle), cos(currentStartAngle));
     Point cp1 = currentStartPoint;
     cp1 += tangentStart * kappa;
-    
+
     Point revTangentEnd(sin(currentEndAngle), -cos(currentEndAngle));
     Point cp2 = currentEndPoint;
     cp2 += revTangentEnd * kappa;
-    
+
     aSink->BezierTo(cp1, cp2, currentEndPoint);
-    
-    arcSweepLeft -= M_PI / 2.0f;
+
+    arcSweepLeft -= Float(M_PI / 2.0f);
     currentStartAngle = currentEndAngle;
   }
 }
